Example that breaks dropdown
For some reason, the drop shows "undefined" if the contents of the drop
contains the same element type as the target inside in an ng-repeat.

In this example, the target is an a tag. In the dropdown, we should see
a list of a tags. When you click on any of the items in the list, the
controller should display an alert of the selected item.

If we change the target to a button instead of an a tag, the drop
displays correctly. Likewise, if we change the contents of the list to a
button and leave the target as an a tag, the drop displays correctly. In
both cases where the drop is displayed correctly, the click event on the
list item does not work. I *think* this is because the Scope is lost
once drop.js moves the element to the root of the document.